Show ruby-prof used
ruby-prof showed that originally we were spending almost all our time in
Enumerable#find_all inside the by_product method.  This is because we
were doing a lookup inside an array, which is O(n) time.  I converted
the lookup to be inside a hash, which is O(1) time, so about 8193 times
faster in this case.

                         80.440      0.023      0.000     80.417        8193/8193      FarMar::Product#sales
      99.81%   0.03%     80.440      0.023      0.000     80.417             8193      <Class::FarMar::Sale>#by_product
                         78.779      0.016      0.000     78.763        8193/8193      Enumerable#find_all
                          1.638      0.013      0.000      1.624        8193/8194      <Class::FarMar::Aspect>#all